Backend/haeun/mypage update

+++ backend/src/main/java/com/magicdev/manalgak/domain/user_address/service/command/UserAddressCommand.java
@@ -3,13 +3,12 @@ package com.magicdev.manalgak.domain.user_address.service.command;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 
 @Builder
 @Getter
 public class UserAddressCommand {
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     private String address;
     private String category;
 }
